refactor(pptx): move endpoints out of app.py
Move PPTX request/response models and /api/pptx endpoints into a dedicated plugin module and register them from app startup. Add python-slugify for preset id normalization and document the new module.
