+++ pixi.lock
@@ -1663,7 +1663,7 @@ packages:
 - pypi: ./
   name: canvas-chat
   version: 0.1.84
-  sha256: ab190ae7e30f6138a7edd01de18283c7ff697a9a7e6ddb89329ae0734ad629aa
+  sha256: 8b90ac664bb2fa45f4d01fce02401709abfb009a92e010e01a383125a47d42ba
   requires_dist:
   - fastapi>=0.115.0
   - uvicorn>=0.32.0
@@ -1686,6 +1686,7 @@ packages:
   - youtube-transcript-api>=1.2.3,<2
   - pillow>=12.1.0,<13
   - python-pptx>=1.0.0
+  - python-slugify
   - defusedxml
   - pytest>=8.0.0 ; extra == 'dev'
   - httpx>=0.27.0 ; extra == 'dev'
